LayoutTests/fast/forms/form-control-refresh/button-like-controls-text-color-matches-expected.txt

This tests that buttons, submit buttons, and select buttons all have the same CSS color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.style.color == submit.style.color && submit.style.color == button.style.color is true
PASS successfullyParsed is true

TEST COMPLETE
